RS-5746: Reset state of Saml.cs to original to make a merge possible
